requirements.txt
asgiref==3.5.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-face==0.6.0
azure-common==1.1.27
azure-core==1.21.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.2
Django==4.0.1
django-allauth==0.47.0
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
idna==3.3
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.0
Pillow==9.0.0
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
